Rolled back and fixed carts
Ok, so first of all, I rolled back my code from the lift because it was
not working (I saved it in another branch). Then I fixed some bugs with
the cart control:
-Fixed bug where carts did not stop when you stopped pressing the button
(until they hit the end). They now stop as soon as you release the move
button.
-Fixed bug where we would go directly to error when either end was
pressed, but we were not going forward or backward (we now do nothing,
but can keep doing things if we press the opposite button)